Add tests for the task functionality
